backend/pyproject.toml
[tool.poetry]
name = "pharmaturk-backend"
version = "0.1.0"
description = "MVP бэкенд Turk-Export на Django/DRF с Celery, Redis, PostgreSQL, Swagger, JWT"
authors = ["Turk-Export Team <dev@turk-export.example>"]
readme = "README.md"
packages = [{ include = "config" }, { include = "api" }, { include = "apps" }]

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.6"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.27.2"
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "^2.9.9"
celery = "^5.3.6"
redis = "^5.0.4"
django-environ = "^0.11.2"
django-cors-headers = "^4.4.0"
gunicorn = "^21.2.0"
opensearch-py = "^2.5.0"
django-prometheus = "^2.3.1"
sentry-sdk = {version = "^2.9.0", extras = ["django"]}
python-json-logger = "^2.0.7"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-django = "^4.8.0"
black = "^24.8.0"
isort = "^5.13.2"
flake8 = "^7.1.1"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]


